refactor(agent): agent structure and routing
- Migrate agent routing and middleware setup to dedicated files for better organization
- Implement CORS and authentication middleware in separate functions
- Simplify route configuration in the agent by using a setupRoutes function
- Enhance system information retrieval by distinguishing between virtual and physical interfaces

These changes improve code maintainability and clarity, facilitating future enhancements to the monitoring system.
